django/requirements.txt
aiohttp==3.7.4.post0
aiohttp-sse-client==0.2.0
amqp==5.0.6
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
crc16==0.1.1
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==4.0.1
kombu==5.0.2
mnemonic==0.19
multidict==5.1.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stellar-base-sseclient==0.0.21
stellar-sdk==3.3.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
zipp==3.4.1
